3.12.0 - Large update with a focus on large game performance. Includes many feature requests and some bug fixes.

+++ alembic/versions/eb3f7dc1b16f_add_flag__order_field.py
@@ -5,9 +5,13 @@ Revises: 67bf6ca63a9c
 Create Date: 2018-12-14 10:56:12.295780
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
+from sqlalchemy.engine.reflection import Inspector
 
+conn = op.get_bind()
+inspector = Inspector.from_engine(conn)
+tables = inspector.get_table_names()
 
 # revision identifiers, used by Alembic.
 revision = "eb3f7dc1b16f"
@@ -16,10 +20,26 @@ branch_labels = None
 depends_on = None
 
 
+def _table_has_column(table, column):
+    has_column = False
+    for col in inspector.get_columns(table):
+        if column not in col["name"]:
+            continue
+        has_column = True
+    return has_column
+
+
+def _has_table(table_name):
+    tables = inspector.get_table_names()
+    return table_name in tables
+
+
 def upgrade():
-    op.add_column("flag", sa.Column("_order", sa.INTEGER))
-    op.create_index("order", "flag", ["_order"])
+    if not _table_has_column("flag", "_order"):
+        op.add_column("flag", sa.Column("_order", sa.INTEGER))
+        op.create_index("order", "flag", ["_order"])
 
 
 def downgrade():
-    op.drop_column("flag", "_order")
+    if _table_has_column("flag", "_order"):
+        op.drop_column("flag", "_order")
